examples/langgraph/pyproject.toml
[project]
name = "langgraph-example"
version = "0.1.0"
description = "Build a typed browser agent with LangGraph and Steel"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-anthropic>=0.3.0",
    "langchain-core>=0.3.0",
    "langgraph>=0.5.0",
    "playwright>=1.48.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "steel-sdk>=0.16.0",
]
